% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcca-utils.R
\name{center_mat}
\alias{center_mat}
\title{Row and column center a matrix}
\usage{
center_mat(X)
}
\arguments{
\item{X}{The matrix to center}
}
\value{
(I - 11^T/n) %*% X %*% (I - 11^T/n)
}
\description{
Row and column center a matrix
}

