% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mkcca.R
\name{get_k_matrices}
\alias{get_k_matrices}
\title{Get the kernel matrices associated with each table.}
\usage{
get_k_matrices(x_list, opts)
}
\arguments{
\item{x_list}{A list whose elements are matrices with samples on rows and
features on columns.}

\item{opts}{A list of options specifying how to perform the kcca. See
merge_kernel_opts.}
}
\value{
A list whose i^th element is the n x n kernel matrix associated with
the i^th data set.
}
\description{
Get the kernel matrices associated with each table.
}

