% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcca.R
\name{kcca}
\alias{kcca}
\title{Perform Yamanashi's multiple / integrated kernel canonical correlation
analysis}
\usage{
kcca(x_list, opts)
}
\arguments{
\item{x_list}{A list whose elements are matrices with samples on rows and
features on columns.}

\item{opts}{A list of options specifying how to perform the kcca. See
merge_kernel_opts.}
}
\value{
The eigenvalues and eigenvectors for the eigenproblems (16 / 18) in
the referenced paper.
}
\description{
Both multiple and integrated kccas are handled simultaneously,
by assuming the input data is a list whose i^th element is a list of kernels
to be applied to the i^th data set. The sum of these matrices is used in the
final eigenproblem, as in equation (18) of the paper. However, we can have
multiple attributes for multiple data sets, and each of the i element of the
data list will be expanded into a separate kcca block, as in equation (16).
That is, to do integrated kcca, ensure the $kernels element of opts is a list
of length 2, both of which are lists of kernels to apply to each data set. On
the other hand, to do multiple kcca, ensure that the $kernels element of opts
is a list of length 1 lists, containing the single kernel to apply to that
data set.
}
\examples{
# integrated kcca
x_list <- replicate(2, matrix(rnorm(100), 10, 10), simplify = F)
rbf_kernels <- list(rbfdot(1), rbfdot(10))
opts <- list(kernels = replicate(2, rbf_kernels, simplify = F))
kcca(x_list, opts)

# multiple kcca
x_list <- replicate(10, matrix(rnorm(100), 10, 10), simplify = F)
opts <- list(kernels = replicate(10, list(rbfdot(1)), simplify = F))
kcca(x_list, opts)
}
\references{
"Extracting of correlated gene clusters from multiple genomic
data by generalized kernel canonical correlation analysis."
}

