% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcca-utils.R
\name{merge_kernel_opts}
\alias{merge_kernel_opts}
\title{Merge default options for kernel cca}
\usage{
merge_kernel_opts(opts = list(), p = 1)
}
\arguments{
\item{opts}{A partially specified list of options to use in kcca. The current
possible options are
  $kernels A list of kernlab kernel class objects. Defaults to a list of RBF
   kernels with bandwidth 1.
  $lambdas The regularization parameters for each table. Defaults to 1.}

\item{p}{The number of tables in the kernel CCA.}
}
\value{
The original opts, but with unspecified options filled in with
defaults.
}
\description{
Merge default options for kernel cca
}

