% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mkcca.R
\name{mkcca_eigen_a}
\alias{mkcca_eigen_a}
\title{Get the left hand side matrix in equation (16) of Yamanashi's kcca
method}
\usage{
mkcca_eigen_a(k_list, opts)
}
\arguments{
\item{k_list}{A list whose i^th element is the n x n kernel matrix associated
with the i^th data set.}

\item{opts}{A list of options specifying how to perform the kcca. See
merge_kernel_opts.}
}
\value{
The the left hand side matrix in equation (16) of Yamanashi's kcca
method.
}
\description{
Get the left hand side matrix in equation (16) of Yamanashi's kcca
method
}
\references{
"Extracting of correlated gene clusters from multiple genomic
data by generalized kernel canonical correlation analysis."
}

