% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ikcca.R, R/mkcca.R
\name{mkcca_eigen_b}
\alias{mkcca_eigen_b}
\title{Get the right hand side matrix in equation (16) of
Yamanashi's kcca method}
\usage{
mkcca_eigen_b(k_list, opts)

mkcca_eigen_b(k_list, opts)
}
\arguments{
\item{k_list}{A list whose i^th element is the n x n kernel matrix associated
with the i^th data set.}

\item{opts}{A list of options specifying how to perform the kcca. See
merge_kernel_opts.}

\item{k_list}{A list whose i^th element is a list of  n x n kernel matrices
associated with the i^th data set. For mkcca, only the first kernel matrix
for each data set will be used.}

\item{opts}{A list of options specifying how to perform the kcca. See
merge_kernel_opts.}
}
\value{
The the right hand side matrix in equation (16) of Yamanashi's kcca
method.

The the right hand side matrix in equation (16) of Yamanashi's kcca
method.
}
\description{
Get the right hand side matrix in equation (16) of
Yamanashi's kcca method

Get the right hand side matrix in equation (16) of
Yamanashi's kcca method
}
\references{
"Extracting of correlated gene clusters from multiple genomic
data by generalized kernel canonical correlation analysis."

"Extracting of correlated gene clusters from multiple genomic
data by generalized kernel canonical correlation analysis."
}

