% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kcca.R
\name{postprocess_res}
\alias{postprocess_res}
\title{Extract KCCA Scores}
\usage{
postprocess_res(eigen_res, k_list)
}
\arguments{
\item{eigen_res}{The result of a call to geigen}

\item{k_list}{A list whose i^th element is a list of  n x n kernel matrices}
}
\value{
A list with the following elements,
  $scores: A list whose i^th element is a matrix of scores for the i^th table. \cr
  $values: The eigenvalues of the problem, in descending order. \cr
  $vectors: The eigenvectors of the KCCA problem (not directly interpretable) \cr
}
\description{
Extract KCCA Scores
}

