% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mkcca.R
\name{mkcca}
\alias{mkcca}
\title{Perform Yamanashi's multiple kernel canonical correlation analysis}
\usage{
mkcca(x_list, opts)
}
\arguments{
\item{opts}{A list of options specifying how to perform the kcca. See
merge_kernel_opts.}

\item{k_list}{A list whose i^th element is the n x n kernel matrix associated
with the i^th data set}
}
\value{
The eigenvalues and eigenvectors for the eigenproblem (16) in the
referenced paper.
}
\description{
Perform Yamanashi's multiple kernel canonical correlation analysis
}
\references{
"Extracting of correlated gene clusters from multiple genomic
data by generalized kernel canonical correlation analysis."
}

